# app-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Notatek",
        "description": "Dokumentacja API dla systemu notatek",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Rejestracja nowego użytkownika",
                "operationId": "c0863bef22b87dded4e00f43a25da91c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "user123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "securepassword"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "securepassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Użytkownik zarejestrowany"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logowanie użytkownika",
                "operationId": "0786998f6014156598117db7dcdb82d0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login_or_email",
                                    "password"
                                ],
                                "properties": {
                                    "login_or_email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "securepassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zalogowano pomyślnie"
                    },
                    "401": {
                        "description": "Nieprawidłowe dane logowania"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Wylogowanie użytkownika",
                "operationId": "a4df9f7f964ae5e3dc802521581c1c18",
                "responses": {
                    "200": {
                        "description": "Wylogowano pomyślnie"
                    },
                    "401": {
                        "description": "Nieautoryzowany"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/friends": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Pobieranie listy znajomych",
                "operationId": "5025b8179fce4c9c04f86721e199da27",
                "responses": {
                    "200": {
                        "description": "Lista znajomych"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Dodawanie znajomego",
                "operationId": "33f8a5fae4af3ae188d4d8f246051401",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "friend_login"
                                ],
                                "properties": {
                                    "friend_login": {
                                        "type": "string",
                                        "example": "friendUser"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Znajomy został dodany"
                    },
                    "400": {
                        "description": "Nie można dodać siebie jako znajomego"
                    },
                    "403": {
                        "description": "Nie można dodać administratora"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/friends/{id}": {
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Usuwanie znajomego",
                "operationId": "af6b2a8851a0c9c864c2a8e21aa196f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID znajomego do usunięcia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Znajomy został usunięty"
                    },
                    "404": {
                        "description": "Nie znaleziono użytkownika"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Wpisz token w formacie (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}